+++ edge-modules/metrics-collector/src/Settings.cs
@@ -22,10 +22,6 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
             string dataCollectionEndpoint,
             string dataCollectionRuleId,
             string dataCollectionStreamName,
-            string aadClientId,
-            string aadTenantId,
-            string aadClientCertificatePath,
-            string aadClientCertificatePassword,
             string endpoints,
             int scrapeFrequencySecs,
             UploadTarget uploadTarget,
@@ -48,14 +44,6 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
                 this.DataCollectionRuleId = Preconditions.CheckNonWhiteSpace(dataCollectionRuleId, nameof(dataCollectionRuleId));
                 this.DataCollectionStreamName = Preconditions.CheckNonWhiteSpace(dataCollectionStreamName, nameof(dataCollectionStreamName));
             }
-
-            // These are all optional. When none are set, DefaultAzureCredential picks up ambient
-            // credentials (workload identity federated token, Azure CLI login, managed identity, etc).
-            // When AadClientCertificatePath is set, a certificate-based app registration is used instead.
-            this.AadClientId = aadClientId;
-            this.AadTenantId = aadTenantId;
-            this.AadClientCertificatePath = aadClientCertificatePath;
-            this.AadClientCertificatePassword = aadClientCertificatePassword;
 
             this.Endpoints = new List<string>();
             foreach (string endpoint in endpoints.Split(","))
@@ -100,10 +88,6 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
                     configuration.GetValue<string>("DataCollectionEndpoint", null),
                     configuration.GetValue<string>("DataCollectionRuleId", null),
                     configuration.GetValue<string>("DataCollectionStreamName", null),
-                    configuration.GetValue<string>("AadClientId", null),
-                    configuration.GetValue<string>("AadTenantId", null),
-                    configuration.GetValue<string>("AadClientCertificatePath", null),
-                    configuration.GetValue<string>("AadClientCertificatePassword", null),
                     configuration.GetValue<string>("MetricsEndpointsCSV", "http://edgeHub:9600/metrics,http://edgeAgent:9600/metrics"),
                     configuration.GetValue<int>("ScrapeFrequencyInSecs", 300),
                     configuration.GetValue<UploadTarget>("UploadTarget", UploadTarget.AzureMonitor),
@@ -137,10 +121,6 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
                 settings.DataCollectionEndpoint,
                 settings.DataCollectionRuleId,
                 settings.DataCollectionStreamName,
-                settings.AadClientId,
-                settings.AadTenantId,
-                settings.AadClientCertificatePath,
-                settings.AadClientCertificatePassword,
                 string.Join(',', settings.Endpoints),
                 settings.ScrapeFrequencySecs,
                 settings.UploadTarget,
@@ -160,14 +140,6 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
         public string DataCollectionRuleId { get; }
 
         public string DataCollectionStreamName { get; }
-
-        public string AadClientId { get; }
-
-        public string AadTenantId { get; }
-
-        public string AadClientCertificatePath { get; }
-
-        public string AadClientCertificatePassword { get; }
 
         public List<string> Endpoints { get; }
 
@@ -201,9 +173,6 @@ namespace Microsoft.Azure.Devices.Edge.Azure.Monitor
                 { nameof(this.DataCollectionEndpoint), this.DataCollectionEndpoint ?? string.Empty },
                 { nameof(this.DataCollectionRuleId), this.DataCollectionRuleId ?? string.Empty },
                 { nameof(this.DataCollectionStreamName), this.DataCollectionStreamName ?? string.Empty },
-                { nameof(this.AadClientId), this.AadClientId ?? string.Empty },
-                { nameof(this.AadTenantId), this.AadTenantId ?? string.Empty },
-                { nameof(this.AadClientCertificatePath), this.AadClientCertificatePath ?? string.Empty },
                 { nameof(this.Endpoints), JsonConvert.SerializeObject(this.Endpoints, Formatting.Indented) },
                 { nameof(this.ScrapeFrequencySecs), this.ScrapeFrequencySecs.ToString() },
                 { nameof(this.UploadTarget), Enum.GetName(typeof(UploadTarget), this.UploadTarget) },